move everything from MainActivity to HomeFragment
